+++ orury-client/src/main/java/org/fastcampus/oruryclient/auth/jwt/JwtTokenFilter.java
@@ -46,7 +46,7 @@ public class JwtTokenFilter extends OncePerRequestFilter {
 
     @Override
     protected boolean shouldNotFilter(HttpServletRequest request) throws ServletException {
-        String[] excludePath = {"/api/v1/auth", "/api/v1/auth/login", "/swagger-ui/index.html", "/swagger-resources", "/favicon.ico"};
+        String[] excludePath = {"/api/v1/auth", "/swagger-ui", "/swagger-resources", "/favicon.ico"};
         String path = request.getRequestURI();
         return Arrays.stream(excludePath).anyMatch(path::startsWith);
     }
