+++ orury-client/src/main/java/org/fastcampus/oruryclient/auth/strategy/KakaoLoginStrategy.java
@@ -12,9 +12,9 @@ import org.fastcampus.oruryclient.auth.jwt.JwtTokenProvider;
 import org.fastcampus.orurycommon.error.code.AuthErrorCode;
 import org.fastcampus.orurycommon.error.exception.AuthException;
 import org.fastcampus.orurydomain.auth.dto.JwtToken;
-import org.fastcampus.orurydomain.auth.dto.KakaoAccountDto;
-import org.fastcampus.orurydomain.auth.dto.KakaoTokenDto;
 import org.fastcampus.orurydomain.auth.dto.LoginDto;
+import org.fastcampus.orurydomain.auth.dto.kakao.KakaoAccountDto;
+import org.fastcampus.orurydomain.auth.dto.kakao.KakaoTokenDto;
 import org.fastcampus.orurydomain.user.db.model.User;
 import org.fastcampus.orurydomain.user.db.repository.UserRepository;
 import org.fastcampus.orurydomain.user.dto.UserDto;
