+++ orury-client/src/main/java/org/fastcampus/oruryclient/global/config/SecurityConfig.java
@@ -45,10 +45,6 @@ public class SecurityConfig {
                 .sessionManagement(sessionManagement ->
                         sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 )
-                .oauth2Login(oauth -> oauth
-                        .userInfoEndpoint(userInfo -> userInfo.userService(oAuth2UserService))
-                        .successHandler(oAuthSuccessHandler)
-                )
                 .build();
     }
