+++ orury-domain/src/main/java/org/fastcampus/orurydomain/auth/dto/kakao/KakaoTokenDto.java
@@ -1,4 +1,4 @@
-package org.fastcampus.orurydomain.auth.dto;
+package org.fastcampus.orurydomain.auth.dto.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
